refactor(driver): move dbms to index module
