+++ jina/drivers/index.py
@@ -3,8 +3,6 @@ __license__ = "Apache-2.0"
 
 from typing import Iterable, Optional
 
-import numpy as np
-
 from . import BaseExecutableDriver, FlatRecursiveMixin
 
 if False:
@@ -68,3 +66,29 @@ class KVIndexDriver(BaseIndexDriver):
             keys, values = zip(*info)
             self.check_key_length(keys)
             self.exec_fn(keys, values)
+
+
+class DBMSIndexDriver(BaseIndexDriver):
+    """Forwards ids, vectors, serialized Document to a BaseDBMSIndexer"""
+
+    def _apply_all(self, docs: 'DocumentSet', *args, **kwargs) -> None:
+        info = [
+            (
+                doc.id,
+                doc.embedding,
+                self._doc_without_embedding(doc).SerializeToString(),
+            )
+            for doc in docs
+        ]
+        if info:
+            ids, vecs, metas = zip(*info)
+            self.check_key_length(ids)
+            self.exec_fn(ids, vecs, metas)
+
+    @staticmethod
+    def _doc_without_embedding(d):
+        from .. import Document
+
+        new_doc = Document(d, copy=True)
+        new_doc.ClearField('embedding')
+        return new_doc
